// ExistingCode/src/data-access/domain/tables/survey/schema.ts
export const SurveyTableInsertBaseSchema = z.object({
	/**
	 * dataType: uniqueidentifier
	 * defaultValue: (newid())
	 */
	surveyId: z.any().optional(),
	/**
	 * dataType: nvarchar
	 */
	organization: z.string().max(30),
	/**
	 * dataType: uniqueidentifier
	 */
	collarId: z.any(),
	/**
	 * dataType: bit
	 * defaultValue: ((0))
	 */
	activeInd: z.any().optional(),
	/**
	 * dataType: uniqueidentifier
	 */
	loggingEventId: z.any(),
	/**
	 * dataType: nvarchar
	 */
	downHoleSurveyMethod: z.string().max(50),
	/**
	 * dataType: datetimeoffset
	 */
	surveyedOnDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	surveyCompany: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	surveyOperator: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	surveyInstrument: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	surveyReliability: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	grid: z.string().max(50).nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	comments: z.string().max(1000).nullable().optional(),
	/**
	 * dataType: bit
	 */
	validation: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	dataSource: z.string().max(255),
	/**
	 * dataType: datetimeoffset
	 * defaultValue: (sysdatetimeoffset())
	 */
	createdOnDt: z.any().optional(),
	/**
	 * dataType: nvarchar
	 * defaultValue: (suser_sname())
	 */
	createdBy: z.string().max(50).optional(),
	/**
	 * dataType: datetimeoffset
	 */
	modifiedOnDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	modifiedBy: z.string().max(50).nullable().optional(),
	/**
	 * dataType: timestamp
	 */
	rv: z.date(),
	/**
	 * dataType: tinyint
	 * defaultValue: ((0))
	 */
	rowStatus: z.number().int().optional(),
});